Python/micron-template/pyproject.toml
[project]
name = "micron-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fire>=0.7.1",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
]

[dependency-groups]
dev = [
    "cx-freeze>=8.4.1",
]
